Move validator package into controller
