+++ controller/validator.go
@@ -1,11 +1,8 @@
-// Package validator provides functionality for validating groups and units.
-package validator
+package controller
 
 import (
 	"sort"
 	"strings"
-
-	"github.com/giantswarm/inago/controller"
 )
 
 // StringsUnique returns true if all strings in the list are unique,
@@ -77,7 +74,7 @@ func StringsHaveOrNot(s []string, c string) bool {
 
 // ValidateRequest takes a Request, and returns whether it is valid or not.
 // If the request is not valid, the error provides more details.
-func ValidateRequest(request controller.Request) (bool, error) {
+func ValidateRequest(request Request) (bool, error) {
 	// Check there are units in the group.
 	if len(request.Units) == 0 {
 		return false, noUnitsInGroupError
@@ -120,7 +117,7 @@ func ValidateRequest(request controller.Request) (bool, error) {
 // ValidateMultipleRequest takes a list of Requests, and returns whether
 // they are valid together or not.
 // If the requests are not valid, the error returned provides more details.
-func ValidateMultipleRequest(requests []controller.Request) (bool, error) {
+func ValidateMultipleRequest(requests []Request) (bool, error) {
 	groupNames := []string{}
 
 	for _, request := range requests {
